created Omniauthable concern
