+++ app/controllers/application_controller.rb
@@ -1,16 +1,17 @@
 class ApplicationController < ActionController::Base
   include Pundit
+  
   before_action :authenticate_user!
   before_action :configure_permitted_parameters, if: :devise_controller?
   skip_forgery_protection
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
   def after_sign_up_path_for(resource)
-    edit_select_diet_path(current_user.username)
+    select_diet_path
   end
 
   def after_sign_in_path_for(resource)
-    edit_select_diet_path(current_user.username)
+    select_diet_path
   end
 
   def configure_permitted_parameters
